Removed old DataManager and Stats
